executor: only call libcontainer-shim on main in linux
